requirements.txt
Jinja2==3.1.6
python-dotenv==1.1.0
requests==2.32.3
